Move the backpropagation code to its own directory
