+++ CMakeLists.txt
@@ -46,26 +46,26 @@ set(SOURCES
   compression/sfp.h
   compression/sfp-inl.h
   compression/test_util.h
+  backprop/backward.cc
+  backprop/backward.h
+  backprop/backward-inl.h
+  backprop/backward_scalar.h
+  backprop/common_scalar.cc
+  backprop/common_scalar.h
+  backprop/forward.cc
+  backprop/forward.h
+  backprop/forward-inl.h
+  backprop/forward_scalar.h
+  backprop/optimizer.cc
+  backprop/optimizer.h
   gemma/configs.h
   gemma/activations.cc
   gemma/activations.h
-  gemma/backward.cc
-  gemma/backward.h
-  gemma/backward-inl.h
-  gemma/backward_scalar.h
   gemma/common.h
   gemma/common-inl.h
-  gemma/common_scalar.cc
-  gemma/common_scalar.h
-  gemma/forward.cc
-  gemma/forward.h
-  gemma/forward-inl.h
-  gemma/forward_scalar.h
   gemma/gemma.cc
   gemma/gemma.h
   gemma/ops.h
-  gemma/optimizer.cc
-  gemma/optimizer.h
   gemma/weights.cc
   gemma/weights.h
   util/app.h
@@ -122,11 +122,11 @@ enable_testing()
 include(GoogleTest)
 
 set(GEMMA_TEST_FILES
+  backprop/backward_test.cc
+  backprop/backward_scalar_test.cc
+  backprop/optimize_test.cc
   gemma/ops_test.cc
   gemma/gemma_test.cc
-  gemma/backward_test.cc
-  gemma/backward_scalar_test.cc
-  gemma/optimize_test.cc
 )
 
 foreach (TESTFILE IN LISTS GEMMA_TEST_FILES)
