+++ backprop/backward.cc
@@ -13,17 +13,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gemma/backward.h"
+#include "backprop/backward.h"
 
 // Compiles this file for multiple architectures via "foreach_target.h", to
 // which we pass the filename via macro 'argument'.
 #undef HWY_TARGET_INCLUDE
-#define HWY_TARGET_INCLUDE "gemma/backward.cc"  // NOLINT
+#define HWY_TARGET_INCLUDE "backprop/backward.cc"  // NOLINT
 #include "hwy/foreach_target.h"        // IWYU pragma: keep
 
-#include "gemma/backward-inl.h"
-#include "gemma/weights.h"
 #include "hwy/highway.h"
+// After highway.h
+#include "backprop/backward-inl.h"
+#include "gemma/weights.h"
 
 HWY_BEFORE_NAMESPACE();
 namespace gcpp {
