+++ backprop/forward.cc
@@ -13,16 +13,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gemma/forward.h"
+#include "backprop/forward.h"
 
 // Compiles this file for multiple architectures via "foreach_target.h", to
 // which we pass the filename via macro 'argument'.
 #undef HWY_TARGET_INCLUDE
-#define HWY_TARGET_INCLUDE "gemma/forward.cc"  // NOLINT
+#define HWY_TARGET_INCLUDE "backprop/forward.cc"  // NOLINT
 #include "hwy/foreach_target.h"        // IWYU pragma: keep
 
-#include "hwy/highway.h"  // IWYU pragma: keep
-#include "gemma/forward-inl.h"
+#include "hwy/highway.h"
+// After highway.h
+#include "backprop/forward-inl.h"
 #include "gemma/weights.h"
 
 HWY_BEFORE_NAMESPACE();
