+++ backprop/optimizer.cc
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gemma/optimizer.h"
+#include "backprop/optimizer.h"
 
 #include <random>
